+++ src/Ortographic.cpp
@@ -8,9 +8,9 @@ namespace pm {
 void Ortographic::renderScene(World &world, RGBColor *frame, int startX, int startY, int tileWidth, int tileHeight)
 {
 	Ray ray;
-	ray.d = Vector3(0.0, 0.0, -1.0);
+	ray.d = Vector3(0.0f, 0.0f, -1.0f);
 	const ViewPlane &vp = world.viewPlane();
-	const float zw = 100.0; // hard-coded
+	const float zw = 100.0f; // hard-coded
 
 	const int numSamples = vp.samplerState().numSamples();
 	for (int r = startY; r < startY + tileHeight; r++)
@@ -21,12 +21,12 @@ void Ortographic::renderScene(World &world, RGBColor *frame, int startX, int sta
 			for (int j = 0; j < numSamples; j++)
 			{
 				const Vector2 sp = vp.samplerState().sampleUnitSquare();
-				const float x = vp.pixelSize() * (c - 0.5 * vp.width() + sp.x);
-				const float y = vp.pixelSize() * (r - 0.5 * vp.height() + sp.y);
+				const float x = vp.pixelSize() * (c - 0.5f * vp.width() + sp.x);
+				const float y = vp.pixelSize() * (r - 0.5f * vp.height() + sp.y);
 				ray.o = Vector3(x, y, zw);
 				pixel += world.tracer().traceRay(ray, 0);
 			}
-			pixel /= numSamples;
+			pixel /= static_cast<float>(numSamples);
 			frame[r * vp.width() + c] = pixel;
 		}
 	}
